requirements.txt
pandas==2.3.0
python-dotenv==1.1.1
selenium==4.34.0
webdriver_manager==4.0.2
